teeeeeeschke's real fixes

+++ src/eshop/Datenstrukturen/Massengutartikel.java
@@ -4,7 +4,7 @@ public class Massengutartikel extends Artikel{
 
 	private int packungsgroesse;
 
-	public Massengutartikel(String bez, int nr, int preis, int bestand, int packungsgroesse) {
+	public Massengutartikel(String bez, int nr, float preis, int bestand, int packungsgroesse) {
 		super(bez, nr, preis, bestand);
 		this.setPackungsgroesse(packungsgroesse);
 	}
